file structure is ready
