+++ js/lib/calculationOfResults.test.js
