+++ js/lib/countPoints.js
