+++ js/lib/countPoints.test.js
