+++ js/lib/setStateGame.js
@@ -1,4 +1,4 @@
-import {initialState, stateGame} from './data';
+import {initialState, stateGame} from '../data/data';
 
 const setStateGame = () => {
   for (const key in initialState) {
