+++ js/lib/timer.js
