+++ js/screens/welcome/welcome.js
@@ -1,11 +1,9 @@
-import WelcomeView from './welcome-view';
-import {displayScreen} from '../functions/screenRender';
+import WelcomeView from '../../views/welcome-view';
+import {displayScreen} from '../../lib/screenRender';
 import showGame from '../game/game';
-import setStateGame from '../setStateGame';
 
 const welcome = new WelcomeView();
 welcome.startHandler = () => {
-  setStateGame();
   const game = showGame();
   displayScreen(game.element);
   game.init();
