+++ js/views/game-view.js
@@ -1,4 +1,4 @@
-import AbstractView from '../../abstract-view';
+import AbstractView from './abstract-view';
 
 class GameView extends AbstractView {
   constructor() {
