+++ js/views/level-view.js
@@ -1,4 +1,4 @@
-import AbstractView from '../../abstract-view';
+import AbstractView from './abstract-view';
 
 class LevelView extends AbstractView {
   constructor(level) {
